% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesCPs-package.R
\docType{package}
\name{BayesChange-package}
\alias{BayesChange}
\alias{BayesChange-package}
\title{BayesChange: Tools for Bayesian Change Points Analysis}
\description{
This package containts functions for performing Change Points Detection both on univariate and multivariate time series. It also contains functions to cluster time series (either univariate or multivariate) and survival functions with common change points.
}
\author{
\strong{Maintainer}: Luca Danese \email{danese.luca1@gmail.com}

}
\keyword{internal}
