% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustCpObj.R
\name{ClustCpObj}
\alias{ClustCpObj}
\title{ClustCpObj class constructor}
\usage{
ClustCpObj(
  data = NULL,
  n_iterations = NULL,
  n_burnin = NULL,
  clust = NULL,
  orders = NULL,
  time = NULL,
  lkl = NULL,
  norm_vec = NULL,
  rho = NULL,
  kernel_ts = NULL,
  kernel_epi = NULL,
  univariate_ts = NULL
)
}
\arguments{
\item{data}{a vector or a matrix containing the values of the time series;}

\item{n_iterations}{number of iterations of the MCMC algorithm;}

\item{n_burnin}{number of MCMC iterations to exclude in the posterior estimate;}

\item{orders}{a matrix where each row corresponds to the output order of the corresponding iteration;}

\item{time}{computational time in seconds;}

\item{univariate_ts}{TRUE/FALSE if time series is univariate or not;}

\item{gamma_MCMC}{traceplot for \eqn{\gamma};}

\item{gamma_MCMC_01}{a \eqn{0/1} vector, the \eqn{n}-th element is equal to \eqn{1} if the proposed \eqn{\gamma} was accepted, \eqn{0} otherwise;}

\item{sigma_MCMC}{traceplot for \eqn{\sigma};}

\item{sigma_MCMC_01}{a \eqn{0/1} vector, the \eqn{n}-th element is equal to \eqn{1} if the proposed \eqn{\sigma} was accepted, \eqn{0} otherwise;}

\item{theta_MCMC}{traceplot for \eqn{\theta};}
}
\description{
A constructor for the \code{ClustCpObj} class. The class \code{ClustCpObj} contains...
}
\examples{

## put example here

}
