% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cluster_cp_uni}
\alias{cluster_cp_uni}
\title{Clustering univariate times series with common changes in time}
\usage{
cluster_cp_uni(
  data,
  n_iterations,
  B,
  L,
  gamma,
  a = 1,
  b = 1,
  c = 1,
  q = 0.5,
  alpha_SM = 0.1,
  user_seed = 1234L
)
}
\arguments{
\item{data}{a matrix where each row is an observation and each column corresponds to a discrete time.}

\item{n_iterations}{number of MCMC iterations.}

\item{B}{number of orders for the normalisation constant.}

\item{L}{number of split-merge steps for the proposal step.}

\item{gamma, a, b, c}{parameters \eqn{\gamma} for the integrated likelihood.}

\item{q}{probability of a split in the split-merge proposal and acceleration step.}

\item{alpha_SM}{\eqn{\alpha} for the split-merge proposal and acceleration step.}

\item{user_seed}{seed for random distribution generation.}
}
\value{
TO DO
}
\description{
Clustering univariate times series with common changes in time
}
