% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ClusteringCPsEPI}
\alias{ClusteringCPsEPI}
\title{Clustering Epidemiological survival functions with common changes in time}
\usage{
ClusteringCPsEPI(
  data,
  niter,
  M,
  B,
  L,
  gamma = 1/8,
  alpha = 1,
  q = 0.1,
  dt = 0.1,
  a0 = 4,
  b0 = 10,
  c0 = 1,
  d0 = 1,
  MH_var = 0.01,
  S0 = 1,
  R0 = 0,
  p = 0.003,
  coarsening = 1,
  user_seed = 1234L
)
}
\arguments{
\item{data}{First value}

\item{niter}{Second value}
}
\value{
TO DO
}
\description{
Clustering Epidemiological survival functions with common changes in time
}
