% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ClusteringCPsMultivariateTS}
\alias{ClusteringCPsMultivariateTS}
\title{Clustering multivariate times series with common changes in time}
\usage{
ClusteringCPsMultivariateTS(
  data,
  n_iterations,
  B,
  L,
  gamma,
  k_0,
  nu_0,
  phi_0,
  m_0,
  q = 0.5,
  alpha_SM = 0.1,
  user_seed = 1234L
)
}
\arguments{
\item{data}{First value}

\item{n_iterations}{Second value}

\item{B}{prova}

\item{L}{prova}

\item{gamma}{prova}

\item{k_0}{prova}

\item{nu_0}{prova}

\item{phi_0}{prova}

\item{m_0}{prova}

\item{q}{prova}

\item{alpha_SM}{prova}

\item{user_seed}{prova}
}
\value{
TO DO
}
\description{
Clustering multivariate times series with common changes in time
}
