% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ClusteringCPsUnivariateTS}
\alias{ClusteringCPsUnivariateTS}
\title{Clustering univariate times series with common changes in time}
\usage{
ClusteringCPsUnivariateTS(
  data,
  n_iterations,
  B,
  L,
  gamma,
  a = 1,
  b = 1,
  c = 1,
  q = 0.5,
  alpha_SM = 0.1,
  user_seed = 1234L
)
}
\arguments{
\item{data}{First value}

\item{n_iterations}{Second value}
}
\value{
TO DO
}
\description{
Clustering univariate times series with common changes in time
}
