% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetectCpObj.R
\name{DetectCpObj}
\alias{DetectCpObj}
\title{DetectCpObj class constructor}
\usage{
DetectCpObj(
  data = NULL,
  n_iterations = NULL,
  n_burnin = NULL,
  orders = NULL,
  time = NULL,
  entropy_MCMC = NULL,
  lkl_MCMC = NULL,
  phi_MCMC = NULL,
  sigma_MCMC = NULL,
  delta_MCMC = NULL,
  I0_MCMC = NULL,
  kernel_ts = NULL,
  kernel_epi = NULL,
  univariate_ts = NULL
)
}
\arguments{
\item{data}{a vector or a matrix containing the values of the time series;}

\item{n_iterations}{number of iterations of the MCMC algorithm;}

\item{n_burnin}{number of MCMC iterations to exclude in the posterior estimate;}

\item{orders}{a matrix where each row corresponds to the output order of the corresponding iteration;}

\item{time}{computational time in seconds;}

\item{phi_MCMC}{traceplot for \eqn{\gamma}.}

\item{sigma_MCMC}{traceplot for \eqn{\sigma}.}

\item{delta_MCMC}{traceplot for \eqn{\delta}.}

\item{I0_MCMC}{traceplot for \eqn{I_0}.}

\item{kernel_ts}{if TRUE data are time series.}

\item{kernel_epi}{if TRUE data are epidemic diffusions.}

\item{univariate_ts}{TRUE/FALSE if time series is univariate or not.}
}
\description{
A constructor for the \code{DetectCpObj} class. The class \code{DetectCpObj} contains...
}
