% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DetectCPsMultivariateTS}
\alias{DetectCPsMultivariateTS}
\title{Detect Change Points on multivariate time series}
\usage{
DetectCPsMultivariateTS(
  data,
  n_iterations,
  q,
  k_0,
  nu_0,
  phi_0,
  m_0,
  user_seed = 1234L,
  prior_theta_c = 1,
  prior_theta_d = 1,
  prior_var_gamma = 0.1
)
}
\arguments{
\item{data}{First value}

\item{n_iterations}{Second value}
}
\value{
TO DO
}
\description{
Detect Change Points on multivariate time series
}
