% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_clust_VI}
\alias{get_clust_VI}
\title{Estimate order}
\usage{
get_clust_VI(orders_mat)
}
\arguments{
\item{orders_mat}{A matrix where each row corresponds to the output cluster of the corresponding iteration.}
}
\value{
Function \code{get_clust_VI} returns a point estimate for the clustering of the data.
}
\description{
Estimate order
}
