% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustCpObj.R
\name{plot.ClustCpObj}
\alias{plot.ClustCpObj}
\title{Plot estimated partition}
\usage{
\method{plot}{ClustCpObj}(
  x,
  y = NULL,
  loss = "VI",
  maxNClusters = 0,
  nRuns = 16,
  maxZealousAttempts = 10,
  ...
)
}
\arguments{
\item{x}{an object of class \code{ClustCpObj}.}

\item{y}{parameter of the generic method.}

\item{loss}{The loss function used to estimate the final partition, it can be "VI", "binder", "omARI", "NVI", "ID", "NID".}

\item{maxNClusters}{maximum number of clusters in salso procedure.}

\item{nRuns}{number of runs in salso procedure.}

\item{maxZealousAttempts}{maximum number of zealous attempts in salso procedure.}

\item{...}{parameter of the generic method.}
}
\value{
The function returns a ggplot object representing the time series or the epidemic diffusions colored according to the final partition.
}
\description{
The \code{plot} method plots the estimates partition through the salso algorithm, for a \code{ClustCpObj} class object.
}
\examples{

\donttest{
## Time series

data("stock_uni")

params_uni <- list(a = 1,
                   b = 1,
                   c = 1,
                   phi = 0.1)

out <- clust_cp(data = stock_uni[1:5,], n_iterations = 7500, n_burnin = 2500,
                L = 1, q = 0.5, B = 10000, params = params_uni, kernel = "ts")

plot(out)

## Epidemic diffusions

data("epi_synthetic_multi")

params_epi <- list(M = 1000, xi = 1/8,
                   alpha_SM = 1,
                   a0 = 4,
                   b0 = 10,
                   I0_var = 0.1,
                   avg_blk = 2)

out <- clust_cp(epi_synthetic_multi, n_iterations = 5000, n_burnin = 2000,
                L = 1, B = 1000, params = params_epi, kernel = "epi")

plot(out)

}

}
