% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetectCpObj.R
\name{posterior_estimate.DetectCpObj}
\alias{posterior_estimate.DetectCpObj}
\title{Estimate the change points of the data}
\usage{
\method{posterior_estimate}{DetectCpObj}(
  object,
  show_cp = FALSE,
  loss = "VI",
  maxNClusters = 0,
  nRuns = 16,
  maxZealousAttempts = 10,
  ...
)
}
\arguments{
\item{object}{An object of class \code{DetectCpObj}.}

\item{show_cp}{Logical; if TRUE show change point locations; FALSE show cluster assignment.}

\item{loss}{The loss function used to estimate the final partition, it can be "VI", "binder", "omARI", "NVI", "ID", "NID".}

\item{maxNClusters}{Maximum number of clusters in salso procedure.}

\item{nRuns}{Number of runs in salso procedure.}

\item{maxZealousAttempts}{Maximum number of zealous attempts in salso procedure.}

\item{...}{parameter of the generic method.}
}
\value{
The function returns a vector with the cluster assignment of each observation.
}
\description{
The \code{posterior_estimate} method estimates the change points of the data making use of the salso algorithm, for a \code{DetectCPObj} class object.
}
\examples{


data("eu_inflation")

params_uni <- list(a = 1, b = 1, c = 1, prior_var_phi = 0.1,
                   prior_delta_c = 1, prior_delta_d = 1)

out <- detect_cp(data = eu_inflation[1,], n_iterations = 1000,
                 n_burnin = 100, q = 0.5, params = params_uni,
                 kernel = "ts")

posterior_estimate(out)

}
\references{
D. B. Dahl, D. J. Johnson, and P. Müller (2022), Search Algorithms and Loss
Functions for Bayesian Clustering, \emph{Journal of Computational and
Graphical Statistics}, 31(4), 1189-1201, \doi{10.1080/10618600.2022.2069779}.
}
