% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetectCpObj.R
\name{print.DetectCpObj}
\alias{print.DetectCpObj}
\title{DetectCpObj print method}
\usage{
\method{print}{DetectCpObj}(x, ...)
}
\arguments{
\item{x}{an object of class \code{DetectCpObj}.}

\item{...}{parameter of the generic method.}
}
\description{
The \code{DetectCpObj} method prints which algorithm was run.
}
\examples{

data("eu_inflation")

params_uni <- list(a = 1, b = 1, c = 1, prior_var_phi = 0.1,
                   prior_delta_c = 1, prior_delta_d = 1)

out <- detect_cp(data = eu_inflation[1,], n_iterations = 1000,
                 n_burnin = 100, q = 0.5, params = params_uni,
                 kernel = "ts")
print(out)

}
