% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{psm}
\alias{psm}
\title{Compute the posterior similarity matrix}
\usage{
psm(M)
}
\arguments{
\item{M}{A matrix where each row corresponds to the output cluster of the corresponding iteration.}
}
\value{
Function \code{psm} returns an \eqn{n}\eqn{\times}\eqn{n} posterior similarity matrix.
}
\description{
Compute the posterior similarity matrix
}
