% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SimEpiData}
\alias{SimEpiData}
\title{Simulate epidemiological data}
\usage{
SimEpiData(S0, I0, MaxTime, beta_vec, gamma_0, user_seed = 1234L)
}
\arguments{
\item{S0}{First value}

\item{I0}{Second value}

\item{MaxTime}{Prova}

\item{beta_vec}{Prova}

\item{gamma_0}{Prova}

\item{user_seed}{Prova}
}
\value{
TO DO
}
\description{
Simulate epidemiological data
}
