% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustCpObj.R
\name{summary.ClustCpObj}
\alias{summary.ClustCpObj}
\title{ClustCpObj summary method}
\usage{
\method{summary}{ClustCpObj}(object, ...)
}
\arguments{
\item{object}{an object of class \code{ClustCpObj}.}

\item{...}{parameter of the generic method.}
}
\description{
The \code{ClustCpObj} method returns a summary of the algorithm.
}
\examples{

data("stock_uni")

params_uni <- list(a = 1,
                   b = 1,
                   c = 1,
                   phi = 0.1)

out <- clust_cp(data = stock_uni[1:3,], n_iterations = 1000, n_burnin = 100,
                L = 1, q = 0.5, B = 500, params = params_uni, kernel = "ts")

summary(out)

}
