% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetectCpObj.R
\name{summary.DetectCpObj}
\alias{summary.DetectCpObj}
\title{DetectCpObj summary method}
\usage{
\method{summary}{DetectCpObj}(object, ...)
}
\arguments{
\item{object}{an object of class \code{DetectCpObj};}

\item{...}{parameter of the generic method.}
}
\description{
The \code{DetectCpObj} method returns a summary of the algorithm.
}
\examples{

data("eu_inflation")

params_uni <- list(a = 1, b = 1, c = 1, prior_var_phi = 0.1,
                   prior_delta_c = 1, prior_delta_d = 1)

out <- detect_cp(data = eu_inflation[1,], n_iterations = 1000,
                 n_burnin = 100, q = 0.5, params = params_uni,
                 kernel = "ts")
summary(out)

}
