% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesChange-package.R
\docType{package}
\name{BayesChange-package}
\alias{BayesChange}
\alias{BayesChange-package}
\title{BayesChange: Bayesian Methods for Change Points Analysis}
\description{
Perform change points detection on univariate and multivariate time series and clustering of different types of time dependent data with common change points.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lucadanese/BayesChange}
  \item Report bugs at \url{https://github.com/lucadanese/BayesChange/issues}
}

}
\author{
\strong{Maintainer}: Luca Danese \email{l.danese1@campus.unimib.com} (\href{https://orcid.org/0000-0001-8444-8563}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Riccardo Corradin
  \item Andrea Ongaro
}

}
\keyword{internal}
