% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesChange-package.R
\docType{package}
\name{BayesChange-package}
\alias{BayesChange}
\alias{BayesChange-package}
\title{BayesChange: Bayesian methods for Change Points Analysis.}
\description{
This package containts functions for performing Change Points Detection both on univariate and multivariate time series. It also contains functions to cluster time series (either univariate or multivariate) and survival functions with common change points.
}
\author{
\strong{Maintainer}: Luca Danese \email{l.danese1@campus.unimib.com} (\href{https://orcid.org/0000-0001-8444-8563}{ORCID})

Authors:
\itemize{
  \item Riccardo Corradin
  \item Andrea Ongaro
}

}
\keyword{internal}
