% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesChange-package.R
\docType{package}
\name{BayesChange-package}
\alias{BayesChange}
\alias{BayesChange-package}
\title{BayesChange: Bayesian Methods for Change Points Analysis}
\description{
Perform change points detection on univariate and multivariate time series according to the methods presented by Asael Fabian Martínez and Ramsés H. Mena (2014) \doi{10.1214/14-BA878} and Corradin, Danese and Ongaro (2022) \doi{10.1016/j.ijar.2021.12.019}. It also clusters different types of time dependent data with common change points, see "Model-based clustering of time-dependent observations with common structural changes" (Corradin,Danese,KhudaBukhsh and Ongaro, 2024) \doi{10.48550/arXiv.2410.09552} for details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lucadanese/BayesChange}
  \item Report bugs at \url{https://github.com/lucadanese/BayesChange/issues}
}

}
\author{
\strong{Maintainer}: Luca Danese \email{l.danese1@campus.unimib.it} (\href{https://orcid.org/0000-0001-8444-8563}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Riccardo Corradin
  \item Andrea Ongaro
}

}
\keyword{internal}
