% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesChange-package.R
\docType{package}
\name{BayesChange-package}
\alias{BayesChange}
\alias{BayesChange-package}
\title{BayesChange: Bayesian Methods for Change Point Analysis}
\description{
Performs change point detection on univariate and multivariate time series (Martínez & Mena, 2014, \doi{10.1214/14-BA878} ; Corradin, Danese & Ongaro, 2022, \doi{10.1016/j.ijar.2021.12.019}) and clusters time-dependent data with common change points (Corradin, Danese, KhudaBukhsh & Ongaro, 2026, \doi{10.1007/s11222-025-10756-x}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lucadanese/BayesChange}
  \item Report bugs at \url{https://github.com/lucadanese/BayesChange/issues}
}

}
\author{
\strong{Maintainer}: Luca Danese \email{l.danese1@campus.unimib.it} (\href{https://orcid.org/0000-0001-8444-8563}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Riccardo Corradin
  \item Andrea Ongaro
}

}
\keyword{internal}
