% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetectCpObj.R
\name{DetectCpObj}
\alias{DetectCpObj}
\title{DetectCpObj class constructor}
\usage{
DetectCpObj(
  data = NULL,
  n_iterations = NULL,
  n_burnin = NULL,
  orders = NULL,
  time = NULL,
  gamma_MCMC = NULL,
  gamma_MCMC_01 = NULL,
  sigma_MCMC = NULL,
  sigma_MCMC_01 = NULL,
  theta_MCMC = NULL,
  univariate_ts = NULL
)
}
\arguments{
\item{data}{a vector or a matrix containing the values of the time series;}

\item{n_iterations}{number of iterations of the MCMC algorithm;}

\item{n_burnin}{number of MCMC iterations to exclude in the posterior estimate;}

\item{orders}{a matrix where each row corresponds to the output order of the corresponding iteration;}

\item{time}{computational time in seconds;}

\item{gamma_MCMC}{traceplot for \eqn{\gamma};}

\item{gamma_MCMC_01}{a \eqn{0/1} vector, the \eqn{n}-th element is equal to \eqn{1} if the proposed \eqn{\gamma} was accepted, \eqn{0} otherwise;}

\item{sigma_MCMC}{traceplot for \eqn{\sigma};}

\item{sigma_MCMC_01}{a \eqn{0/1} vector, the \eqn{n}-th element is equal to \eqn{1} if the proposed \eqn{\sigma} was accepted, \eqn{0} otherwise;}

\item{theta_MCMC}{traceplot for \eqn{\theta};}

\item{univariate_ts}{TRUE/FALSE if time series is univariate or not;}
}
\description{
A constructor for the \code{DetectCpObj} class. The class \code{DetectCpObj} contains...
}
