% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DetectCPsUnivariateTS}
\alias{DetectCPsUnivariateTS}
\title{Detect Change Points on univariate time series}
\usage{
DetectCPsUnivariateTS(
  data,
  n_iterations,
  q,
  phi,
  a,
  b,
  c,
  prior_theta_c = 1,
  prior_theta_d = 1,
  user_seed = 1234L
)
}
\arguments{
\item{n_iterations}{number of MCMC iterations.}

\item{q}{probability of performing a split at each iterations.}

\item{phi}{parameter \eqn{\phi} of the integrated likelihood function.}

\item{a}{prova}

\item{b}{prova}

\item{c}{prova}

\item{prior_theta_c}{prova}

\item{prior_theta_d}{prova}

\item{user_seed}{prova}

\item{data_vec}{vector of observations.}
}
\value{
TO DO
}
\description{
Detect Change Points on univariate time series
}
