% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DetectCPsUnivariateTS}
\alias{DetectCPsUnivariateTS}
\title{Detect Change Points on univariate time series}
\usage{
DetectCPsUnivariateTS(
  data_vec,
  n_iterations,
  q,
  phi,
  a,
  b,
  c,
  user_seed = 1234L,
  prior_theta_c = 1,
  prior_theta_d = 1
)
}
\arguments{
\item{data_vec}{First value}

\item{n_iterations}{Second value}
}
\value{
TO DO
}
\description{
Detect Change Points on univariate time series
}
