% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustCpObj.R, R/DetectCpObj.R
\name{posterior_estimate}
\alias{posterior_estimate}
\alias{plot_psm}
\title{set generic}
\usage{
plot_psm(object, ...)

posterior_estimate(object, ...)
}
\description{
set generic

set generic
}
\keyword{internal}
