% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_epi_data}
\alias{sim_epi_data}
\title{Simulate epidemiological data}
\usage{
sim_epi_data(S0, I0, max_time, beta_vec, xi_0, user_seed = 1234L)
}
\arguments{
\item{S0}{number of individuals in the population.}

\item{I0}{number of infected individuals at time 0.}

\item{max_time}{maximum observed time.}

\item{beta_vec}{vector with the infection rate for each discrete time.}

\item{xi_0}{the recovery rate of the population, must be in \eqn{(0,1)}.}

\item{user_seed}{seed for random distribution generation.}
}
\value{
Function \code{sim_epi_data} returns a vector with the simulated infection times.
}
\description{
Simulate epidemiological data
}
\examples{

betas <- c(rep(0.45, 25),rep(0.14,25))

inf_times <- as.numeric()

inf_times <- sim_epi_data(10000, 10, 50, betas, 1/8)

}
