% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SimEpiData}
\alias{SimEpiData}
\title{Detect Change Points on univariate time series}
\usage{
SimEpiData(S0, I0, MaxTime, beta_vec, gamma_0, user_seed = 1234L)
}
\arguments{
\item{S0}{First value}

\item{I0}{Second value}
}
\value{
TO DO
}
\description{
Detect Change Points on univariate time series
}
